refactor(home): widget 분리, lint 적용
