+++ wabiz/client/wabiz/lib/core/core.dart
@@ -1,3 +1,8 @@
+// packages
+export 'package:gap/gap.dart';
+export 'package:go_router/go_router.dart';
+
+// app
 export 'colors.dart';
 export 'router.dart';
 export 'theme.dart';
