+++ wabiz/client/wabiz/lib/views/common/bottom_nav_shell.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 import 'package:wabiz/core/core.dart';
 
 class BottomNavShell extends StatelessWidget {
