+++ wabiz/client/wabiz/lib/views/home/widget/home_card_list.dart
@@ -1,0 +1,103 @@
+import 'package:flutter/material.dart';
+import 'package:wabiz/core/core.dart';
+
+class HomeCardList extends StatelessWidget {
+  const HomeCardList({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Expanded(
+      child: ListView.separated(
+        itemCount: 10,
+        separatorBuilder: (context, index) => const Divider(
+          height: 20,
+          color: Colors.transparent,
+        ),
+        itemBuilder: (context, index) {
+          if (index == 0) return const SizedBox.shrink();
+          return const HomeCard();
+        },
+      ),
+    );
+  }
+}
+
+class HomeCard extends StatelessWidget {
+  const HomeCard({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      child: Container(
+        margin: const EdgeInsets.symmetric(horizontal: 16),
+        decoration: BoxDecoration(
+          color: AppColors.white,
+          borderRadius: BorderRadius.circular(10),
+          boxShadow: [
+            BoxShadow(
+              offset: const Offset(0, 8),
+              color: AppColors.black.withOpacity(.1),
+              blurRadius: 30,
+              spreadRadius: 4,
+            ),
+          ],
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Container(
+              height: 220,
+              decoration: const BoxDecoration(
+                color: AppColors.wabizGray,
+                borderRadius: BorderRadius.only(
+                  topLeft: Radius.circular(10),
+                  topRight: Radius.circular(10),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(16),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  const Text(
+                    '123122명이 기다려요',
+                    style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.w700,
+                      color: AppColors.primary,
+                    ),
+                  ),
+                  const Gap(8),
+                  const Text('아이돌 관리비법 '),
+                  const Gap(16),
+                  Text(
+                    '세상에 없던 브랜드',
+                    style: TextStyle(
+                      color: AppColors.wabizGray.shade500,
+                    ),
+                  ),
+                  const Gap(16),
+                  Container(
+                    padding:
+                        const EdgeInsets.symmetric(horizontal: 6, vertical: 4),
+                    decoration: BoxDecoration(
+                      color: AppColors.bg,
+                      borderRadius: BorderRadius.circular(3),
+                    ),
+                    child: const Text(
+                      '오픈예정',
+                      style: TextStyle(fontSize: 10),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
