+++ wabiz/client/wabiz/lib/views/home/widget/search_bar.dart
@@ -1,0 +1,54 @@
+import 'package:flutter/material.dart';
+import 'package:wabiz/core/core.dart';
+
+class SearchBar extends StatelessWidget {
+  const SearchBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(16.0),
+      child: Row(
+        children: [
+          Expanded(
+            child: SizedBox(
+              height: 40,
+              child: TextFormField(
+                decoration: InputDecoration(
+                  hintText: '새로운 일상이 필요하신가요?',
+                  suffixIcon: const Icon(Icons.search),
+                  suffixIconColor: AppColors.wabizGray.shade400,
+                  contentPadding: const EdgeInsets.symmetric(
+                    horizontal: 16,
+                  ),
+                  border: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(5),
+                    borderSide: BorderSide(
+                      color: AppColors.wabizGray.shade100,
+                    ),
+                  ),
+                  enabledBorder: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(5),
+                    borderSide: BorderSide(
+                      color: AppColors.wabizGray.shade100,
+                    ),
+                  ),
+                  focusedBorder: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(5),
+                    borderSide: const BorderSide(
+                      color: AppColors.primary,
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+          IconButton(
+            onPressed: () {},
+            icon: const Icon(Icons.notifications_none),
+          ),
+        ],
+      ),
+    );
+  }
+}
